refactor(runner): isolate service drain and resume orchestration
